Change high_surrogate to Option<u16>

+++ src/platform_impl/windows/event_loop.rs
@@ -1035,11 +1035,11 @@ unsafe extern "system" fn public_window_callback<T>(
             let low_surrogate = 0xDC00 <= wparam && wparam <= 0xDFFF;
 
             if high_surrogate {
-                subclass_input.window_state.lock().high_surrogate = Some(wparam);
+                subclass_input.window_state.lock().high_surrogate = Some(wparam as u16);
             } else if low_surrogate {
                 let mut window_state = subclass_input.window_state.lock();
                 if let Some(high_surrogate) = window_state.high_surrogate.take() {
-                    let pair = [high_surrogate as u16, wparam as u16];
+                    let pair = [high_surrogate, wparam as u16];
                     if let Some(Ok(chr)) = std::char::decode_utf16(pair.iter().copied()).next() {
                         subclass_input.send_event(Event::WindowEvent {
                             window_id: RootWindowId(WindowId(window)),
